+++ realm/realm-library/src/main/java/io/realm/RealmConfiguration.java
@@ -243,7 +243,7 @@ public class RealmConfiguration {
      *
      * @return {@code true} if the Realm file exists, {@code false} otherwise.
      */
-    boolean realmExists() {
+    protected boolean realmExists() {
         return new File(canonicalPath).exists();
     }