+++ realm/realm-library/src/objectServer/java/io/realm/mongodb/AppConfiguration.java
@@ -191,6 +191,7 @@ public class AppConfiguration {
      *
      * @return the app name.
      */
+    @Nullable
     public String getAppName() {
         return appName;
     }
@@ -200,6 +201,7 @@ public class AppConfiguration {
      *
      * @return the app version.
      */
+    @Nullable
     public String getAppVersion() {
         return appVersion;
     }
@@ -306,6 +308,45 @@ public class AppConfiguration {
     @Nullable
     public HttpLogObfuscator getHttpLogObfuscator() {
         return httpLogObfuscator;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        AppConfiguration that = (AppConfiguration) o;
+
+        if (requestTimeoutMs != that.requestTimeoutMs) return false;
+        if (!appId.equals(that.appId)) return false;
+        if (appName != null ? !appName.equals(that.appName) : that.appName != null) return false;
+        if (appVersion != null ? !appVersion.equals(that.appVersion) : that.appVersion != null)
+            return false;
+        if (!baseUrl.toString().equals(that.baseUrl.toString())) return false;
+        if (!defaultErrorHandler.equals(that.defaultErrorHandler)) return false;
+        if (!Arrays.equals(encryptionKey, that.encryptionKey)) return false;
+        if (!authorizationHeaderName.equals(that.authorizationHeaderName)) return false;
+        if (!customHeaders.equals(that.customHeaders)) return false;
+        if (!syncRootDir.equals(that.syncRootDir)) return false;
+        if (!codecRegistry.equals(that.codecRegistry)) return false;
+        return httpLogObfuscator != null ? httpLogObfuscator.equals(that.httpLogObfuscator) : that.httpLogObfuscator == null;
+    }
+
+    @Override
+    public int hashCode() {
+        int result = appId.hashCode();
+        result = 31 * result + (appName != null ? appName.hashCode() : 0);
+        result = 31 * result + (appVersion != null ? appVersion.hashCode() : 0);
+        result = 31 * result + baseUrl.toString().hashCode();
+        result = 31 * result + defaultErrorHandler.hashCode();
+        result = 31 * result + Arrays.hashCode(encryptionKey);
+        result = 31 * result + (int) (requestTimeoutMs ^ (requestTimeoutMs >>> 32));
+        result = 31 * result + authorizationHeaderName.hashCode();
+        result = 31 * result + customHeaders.hashCode();
+        result = 31 * result + syncRootDir.hashCode();
+        result = 31 * result + codecRegistry.hashCode();
+        result = 31 * result + (httpLogObfuscator != null ? httpLogObfuscator.hashCode() : 0);
+        return result;
     }
 
     private static Map<String, RegexPatternObfuscator> getLoginObfuscators() {